v0.9.0 Dev Tool - improved deploy script to not require inject config pattern
