+++ Blueprints/command_blueprint.js
@@ -5,13 +5,6 @@ const options = [
 ];
 module.exports = new Command("name", "description", false, false, options);
 
-// imports from files that depend on /Config
-// let ;
-module.exports.injectConfig = function (isProduction) {
-	({} = require("../../helpers.js").injectConfig(isProduction));
-	return this;
-}
-
 module.exports.execute = (interaction) => {
 	// Command specifications go here
 }
