+++ Source/Commands/armory.js
@@ -1,24 +1,12 @@
 const Command = require('../../Classes/Command.js');
+const { embedTemplate } = require('../../helpers.js');
 const { getColor } = require('../elementHelpers.js');
+const { equipmentExists, buildEquipmentDescription, getEquipmentProperty } = require('../equipment/_equipmentDictionary.js');
 
 const options = [
 	{ type: "String", name: "equipment-name", description: "The name of the equipment (case-sensitive)", required: true, choices: {} }
 ];
 module.exports = new Command("armory", "Look up the stats on a type of equipment", false, false, options);
-
-
-let // imports from files that depend on /Config
-	// equipmentDictionary
-	equipmentExists,
-	buildEquipmentDescription,
-	getEquipmentProperty,
-	// helpers
-	embedTemplate;
-module.exports.injectConfig = function (isProduction) {
-	({ equipmentExists, buildEquipmentDescription, getEquipmentProperty } = require('../equipment/_equipmentDictionary.js').injectConfig(isProduction));
-	({ embedTemplate } = require('../../helpers.js').injectConfig(isProduction));
-	return this;
-}
 
 module.exports.execute = (interaction) => {
 	// Command specifications go here
