+++ Source/Commands/feedback.js
@@ -1,16 +1,10 @@
 const Command = require('../../Classes/Command.js');
+const { versionData: { guildId, feedbackChannel } } = require("./../../helpers.js");
 
 const options = [
 	{ type: "String", name: "feedback", description: "Please describe the bug or feature request as clearly as possible", required: true, choices: {} }
 ];
 module.exports = new Command("feedback", "Send PotL feedback to the test server and get an invite", false, false, options);
-
-// imports from files that depend on /Config
-let guildId, feedbackChannel;
-module.exports.injectConfig = function (isProduction) {
-	({ versionData: { guildId, feedbackChannel } } = require("./../../helpers.js").injectConfig(isProduction));
-	return this;
-}
 
 module.exports.execute = (interaction) => {
 	// Post feedback to the test server channel and provide the user an invite
