+++ Source/Commands/give-up.js
@@ -1,14 +1,8 @@
 const Command = require('../../Classes/Command.js');
+const { completeAdventure, getAdventure } = require('../adventureDAO.js');
 
 const options = [];
 module.exports = new Command("give-up", "Lets the adventure leader end the adventure", false, false, options);
-
-// imports from files that depend on /Config
-let completeAdventure, getAdventure;
-module.exports.injectConfig = function (isProduction) {
-	({ completeAdventure, getAdventure } = require('../adventureDAO.js').injectConfig(isProduction));
-	return this;
-}
 
 module.exports.execute = (interaction) => {
 	// Give up on the current adventure
