+++ Source/Commands/ping.js
@@ -1,14 +1,8 @@
 const Command = require('../../Classes/Command.js');
+const { getAdventure } = require('../adventureDAO.js');
 
 const options = [];
 module.exports = new Command("ping", "Remind delvers to input their vote or move", false, false, options);
-
-// imports from files that depend on /Config
-let getAdventure;
-module.exports.injectConfig = function (isProduction) {
-	({ getAdventure } = require('../adventureDAO.js').injectConfig(isProduction));
-	return this;
-}
 
 module.exports.execute = (interaction) => {
 	// Remind delvers to input their vote or move
