+++ Source/Commands/reset.js
@@ -1,18 +1,9 @@
 const Command = require('../../Classes/Command.js');
+const { getGuild } = require('../guildDAO.js');
+const { resetScores } = require('../playerDAO.js');
 
 const options = [];
 module.exports = new Command("reset", "(Manager) Reset player scores for this server", true, false, options);
-
-let // imports from files that depend on /Config
-	// playerDAO
-	resetScores,
-	//guildDAO
-	getGuild;
-module.exports.injectConfig = function (isProduction) {
-	({ resetScores } = require('../playerDAO.js').injectConfig(isProduction));
-	({ getGuild } = require('../guildDAO.js').injectConfig(isProduction));
-	return this;
-}
 
 module.exports.execute = (interaction) => {
 	if (interaction.inGuild()) {
