+++ Source/Enemies/bloodtailhawk.js
@@ -1,20 +1,16 @@
 const Enemy = require("../../Classes/Enemy.js");
+const { dealDamage, addModifier } = require("../combatantDAO.js");
+const { selectRandomFoe, nextRepeat } = require("../enemyDAO.js");
 
-// import from modules that depend on /Config
-let selectRandomFoe, nextRepeat, dealDamage, addModifier;
-module.exports.injectConfig = function (isProduction) {
-	({ selectRandomFoe, nextRepeat } = require("../enemyDAO.js").injectConfig(isProduction));
-	({ dealDamage, addModifier } = require("../combatantDAO.js").injectConfig(isProduction));
-	return new Enemy("Bloodtail Hawk")
-		.setFirstAction("Rake")
-		.addAction({ name: "Rake", effect: rakeEffect, selector: selectRandomFoe, next: nextRepeat })
-		.setBounty(25)
-		.setHp(200)
-		.setSpeed(105)
-		.setElement("Wind")
-		.setStaggerThreshold(1)
-		.setCritDenominator(3);
-}
+module.exports = new Enemy("Bloodtail Hawk")
+	.setFirstAction("Rake")
+	.addAction({ name: "Rake", effect: rakeEffect, selector: selectRandomFoe, next: nextRepeat })
+	.setBounty(25)
+	.setHp(200)
+	.setSpeed(105)
+	.setElement("Wind")
+	.setStaggerThreshold(1)
+	.setCritDenominator(3);
 
 function rakeEffect(target, user, isCrit, adventure) {
 	let damage = 50;
