+++ Source/Enemies/clone.js
@@ -1,12 +1,8 @@
 const Enemy = require("../../Classes/Enemy.js");
 
-// import from modules that depend on /Config
-// let ;
-module.exports.injectConfig = function (isProduction) {
-	return new Enemy("@{clone}")
-		.setBounty(50)
-		.setHp(300)
-		.setSpeed(100)
-		.setElement("@{clone}")
-		.setStaggerThreshold(3);
-}
+module.exports = new Enemy("@{clone}")
+	.setBounty(50)
+	.setHp(300)
+	.setSpeed(100)
+	.setElement("@{clone}")
+	.setStaggerThreshold(3);
