+++ Source/Enemies/geodetortoise.js
@@ -1,21 +1,16 @@
 const Enemy = require("../../Classes/Enemy.js");
+const { addBlock, addModifier, removeModifier, dealDamage } = require("../combatantDAO.js");
+const { selectRandomFoe, selectSelf, nextRandom } = require("../enemyDAO.js");
 
-// import from modules that depend on /Config
-let selectRandomFoe, selectSelf, nextRandom, addBlock, addModifier, removeModifier, dealDamage;
-module.exports.injectConfig = function (isProduction) {
-	({ selectRandomFoe, selectSelf, nextRandom } = require("../enemyDAO.js").injectConfig(isProduction));
-	({ addBlock, addModifier, removeModifier, dealDamage } = require("../combatantDAO.js").injectConfig(isProduction));
-
-	return new Enemy("Geode Tortoise")
-		.setFirstAction("random")
-		.addAction({ name: "Bite", effect: biteEffect, selector: selectRandomFoe, next: nextRandom })
-		.addAction({ name: "Crystallize", effect: crystallizeEffect, selector: selectSelf, next: nextRandom })
-		.setBounty(40)
-		.setHp(350)
-		.setSpeed(85)
-		.setElement("Earth")
-		.setStaggerThreshold(5);
-}
+module.exports = new Enemy("Geode Tortoise")
+	.setFirstAction("random")
+	.addAction({ name: "Bite", effect: biteEffect, selector: selectRandomFoe, next: nextRandom })
+	.addAction({ name: "Crystallize", effect: crystallizeEffect, selector: selectSelf, next: nextRandom })
+	.setBounty(40)
+	.setHp(350)
+	.setSpeed(85)
+	.setElement("Earth")
+	.setStaggerThreshold(5);
 
 function biteEffect(target, user, isCrit, adventure) {
 	let damage = 50;
